# Conflicts:
#	src/main/java/com/codeup/nutritrackr/controllers/DiaryController.java
#	src/main/resources/templates/users/diary.html
#	src/main/resources/templates/users/login.html

+++ src/main/java/com/codeup/nutritrackr/models/Goal.java
@@ -83,6 +83,7 @@ public class Goal {
     }
 
     public int getCalorieGoal() {
+
         return calorieGoal;
     }
 
@@ -109,32 +110,31 @@ public class Goal {
     public int getCarbGoal() {
         return carbGoal;
     }
-
     public void setCarbGoal(int carbGoal) {
         this.carbGoal = carbGoal;
     }
 
-    public double getProteinCaloriesGoal() {
-        return Math.ceil(calorieGoal * (proteinGoal / 100.0));
+    public int getProteinCaloriesGoal() {
+        return (int) Math.ceil(calorieGoal * (proteinGoal / 100.0));
     }
 
-    public double getProteinGramsGoal() {
-        return Math.ceil(getProteinCaloriesGoal() / PROTEIN_CALORIES_PER_GRAM);
+    public int getProteinGramsGoal() {
+        return (int) Math.ceil(getProteinCaloriesGoal() / PROTEIN_CALORIES_PER_GRAM);
     }
 
-    public double getFatCaloriesGoal() {
-        return Math.ceil(calorieGoal * (fatGoal / 100.0));
+    public int getFatCaloriesGoal() {
+        return (int) Math.ceil(calorieGoal * (fatGoal / 100.0));
     }
 
-    public double getFatGramsGoal() {
-        return Math.ceil(getFatCaloriesGoal() / FAT_CALORIES_PER_GRAM);
+    public int getFatGramsGoal() {
+        return (int) Math.ceil(getFatCaloriesGoal() / FAT_CALORIES_PER_GRAM);
     }
 
-    public double getCarbCaloriesGoal() {
-        return Math.ceil(calorieGoal * (carbGoal / 100.0));
+    public int getCarbCaloriesGoal() {
+        return (int) Math.ceil(calorieGoal * (carbGoal / 100.0));
     }
 
-    public double getCarbGramsGoal() {
-        return Math.ceil(getCarbCaloriesGoal() / CARBOHYDRATE_CALORIES_PER_GRAM);
+    public int getCarbGramsGoal() {
+        return (int) Math.ceil(getCarbCaloriesGoal() / CARBOHYDRATE_CALORIES_PER_GRAM);
     }
 }
